docs: fold asset preview player demo back into catalog card stories
